% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyDouble.R
\name{autoplot.FuzzyDouble}
\alias{autoplot.FuzzyDouble}
\title{Plot FuzzyDouble}
\usage{
\method{autoplot}{FuzzyDouble}(object, xlab = "x", ylab = "Double", ...,
  LineCol = "red")
}
\arguments{
\item{object}{The \code{\link{FuzzyDouble}} object}

\item{xlab}{The X-axis label}

\item{ylab}{The Y-axis label}

\item{...}{Extra parameters passed to \code{\link{autoplot}}}

\item{LineCol}{The color of the line representing $y=2x$}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
Plot a FuzzyDouble object with ggplot2
}
\examples{
autoplot(FuzzyDouble(1:10))

}
