% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Double.R
\name{Double}
\alias{Double}
\alias{Double}
\alias{Double.default}
\alias{Double.integer}
\title{Double}
\usage{
Double(x, ...)

\method{Double}{default}(x, ...)

\method{Double}{integer}(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Unused}
}
\value{
A vector containing the double values.
}
\description{
Compute the double value of a vector
}
\details{
The double is calculated by multiplying each value by 2.
}
\examples{
Double(runif(3))

}
