% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0.R
\name{L0}
\alias{L0}
\title{L0}
\usage{
L0(nu, Values)
}
\arguments{
\item{nu}{the fixed points obtained from getuniquesolution}

\item{Values}{The output from getvalue.}
}
\value{
MN and MD.
}
\description{
a function that return MN and MD in MRFM.
}
