% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixH.R
\name{MatrixH}
\alias{MatrixH}
\title{MatrixH}
\usage{
MatrixH(nu, Values)
}
\arguments{
\item{nu}{the fixed points obtained from getuniquesolution}

\item{Values}{The output from getvalue.}
}
\value{
a matrix H.
}
\description{
a function that return matrix H in MRFM.
}
