% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyDouble.R
\name{plot.FuzzyDouble}
\alias{plot.FuzzyDouble}
\title{Plot FuzzyDouble}
\usage{
\method{plot}{FuzzyDouble}(x, xlab = "x", ylab = "Double", ...,
  LineCol = "red")
}
\arguments{
\item{x}{The \code{\link{FuzzyDouble}} object}

\item{xlab}{The X-axis label}

\item{ylab}{The Y-axis label}

\item{...}{Extra parameters passed to \code{\link{plot}}}

\item{LineCol}{The color of the line representing $y=2x$}
}
\description{
Plot a FuzzyDouble object
}
\examples{
plot(FuzzyDouble(1:10))

}
