% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R
\name{sumofMultiples}
\alias{sumofMultiples}
\title{sum of multiples of 3}
\usage{
sumofMultiples(h)
}
\description{
Enter an integer n, then obtain the sum of multiples of 3 from 1 to n
}
\examples{

sumofMultiples(10)
}
